Move to argtable3 argument parser
