+++ config.mk
@@ -8,13 +8,15 @@ PREFIX = /usr
 MANPREFIX = ${PREFIX}/share/man
 DBUS_SERVICES_INSTALL_DIR = ${PREFIX}/share/dbus-1/services
 DBUS_INTERFACES_INSTALL_DIR = ${PREFIX}/share/dbus-1/interfaces
+ARGTABLE3_INCLUDE_DIR = ${PREFIX}/usr/include
+ARGTABLE3_LIB_DIR = ${PREFIX}/usr/lib
 
 X11INC = /usr/X11R6/include
 X11LIB = /usr/X11R6/lib
 
 # includes and libs
-INCS = -I. -I/usr/include -I${X11INC}
-LIBS = -L/usr/lib -lc -lcrypt -L${X11LIB} -lX11 -lXext -lXrandr -lXinerama
+INCS = -I. -I/usr/include -I${X11INC} -I${ARGTABLE3_INCLUDE_DIR}
+LIBS = -L/usr/lib -lc -lcrypt -L${X11LIB} -L${ARGTABLE3_LIB_DIR} -lX11 -lXext -lXrandr -lXinerama -largtable3
 
 # flags
 CPPFLAGS += -DVERSION=\"${VERSION}\" -D_DEFAULT_SOURCE -DHAVE_SHADOW_H
