+++ instantlock.c
@@ -1,558 +1,109 @@
 /* See LICENSE file for license details. */
-#define _XOPEN_SOURCE 500
-#if HAVE_SHADOW_H
-#include <shadow.h>
-#endif
 
-#include <ctype.h>
-#include <errno.h>
-#include <grp.h>
-#include <pwd.h>
-#include <stdarg.h>
-#include <stdlib.h>
 #include <stdio.h>
-#include <string.h>
-#include <unistd.h>
-#include <sys/types.h>
-#include <X11/extensions/Xrandr.h>
-#include <X11/extensions/dpms.h>
-#include <X11/extensions/Xinerama.h>
+#include <argtable3.h>
 
-#include <X11/keysym.h>
-#include <X11/Xlib.h>
-#include <X11/Xutil.h>
-
-#include "arg.h"
-#include "util.h"
-
-char *argv0;
-
-/* global count to prevent repeated error messages */
-int count_error = 0;
-
-enum {
-	INIT,
-	INPUT,
-	FAILED,
-	NUMCOLS
-};
-
-struct lock {
-	int screen;
-	Window root, win;
-	Pixmap pmap;
-	unsigned long colors[NUMCOLS];
-};
-
-struct xrandr {
-	int active;
-	int evbase;
-	int errbase;
-};
-
-#include "config.h"
-
-static void
-die(const char *errstr, ...)
-{
-	va_list ap;
-
-	va_start(ap, errstr);
-	vfprintf(stderr, errstr, ap);
-	va_end(ap);
-	exit(1);
-}
-
-#ifdef __linux__
-#include <fcntl.h>
-#include <linux/oom.h>
-
-static void
-dontkillme(void)
-{
-	FILE *f;
-	const char oomfile[] = "/proc/self/oom_score_adj";
-
-	if (!(f = fopen(oomfile, "w"))) {
-		if (errno == ENOENT)
-			return;
-		die("instantlock: fopen %s: %s\n", oomfile, strerror(errno));
-	}
-	fprintf(f, "%d", OOM_SCORE_ADJ_MIN);
-	if (fclose(f)) {
-		if (errno == EACCES)
-			die("instantlock: unable to disable OOM killer. "
-			    "Make sure to suid or sgid instantlock.\n");
-		else
-			die("instantlock: fclose %s: %s\n", oomfile, strerror(errno));
-	}
-}
-#endif
-
-static void
-writemessage(Display *dpy, Window win, int screen)
-{
-	int len, line_len, width, height, s_width, s_height, i, j, k, tab_replace, tab_size;
-	XGCValues gr_values;
-	XFontStruct *fontinfo;
-	XColor color, dummy;
-	XineramaScreenInfo *xsi;
-	GC gc;
-	fontinfo = XLoadQueryFont(dpy, font_name);
-
-	if (fontinfo == NULL) {
-		if (count_error == 0) {
-			fprintf(stderr, "instantlock: Unable to load font \"%s\"\n", font_name);
-			fprintf(stderr, "instantlock: Try listing fonts with 'instantlock -f'\n");
-			count_error++;
-		}
-		return;
-	}
-
-	tab_size = 8 * XTextWidth(fontinfo, " ", 1);
-
-	XAllocNamedColor(dpy, DefaultColormap(dpy, screen),
-		 text_color, &color, &dummy);
-
-	gr_values.font = fontinfo->fid;
-	gr_values.foreground = color.pixel;
-	gc=XCreateGC(dpy,win,GCFont+GCForeground, &gr_values);
-
-	/*  To prevent "Uninitialized" warnings. */
-	xsi = NULL;
-
-	/*
-	 * Start formatting and drawing text
-	 */
-
-	len = strlen(message);
-
-	/* Max max line length (cut at '\n') */
-	line_len = 0;
-	k = 0;
-	for (i = j = 0; i < len; i++) {
-		if (message[i] == '\n') {
-			if (i - j > line_len)
-				line_len = i - j;
-			k++;
-			i++;
-			j = i;
-		}
-	}
-	/* If there is only one line */
-	if (line_len == 0)
-		line_len = len;
-
-	if (XineramaIsActive(dpy)) {
-		xsi = XineramaQueryScreens(dpy, &i);
-		s_width = xsi[0].width;
-		s_height = xsi[0].height;
-	} else {
-		s_width = DisplayWidth(dpy, screen);
-		s_height = DisplayHeight(dpy, screen);
-	}
-
-	height = s_height*3/7 - (k*20)/3;
-	width  = (s_width - XTextWidth(fontinfo, message, line_len))/2;
-
-	/* Look for '\n' and print the text between them. */
-	for (i = j = k = 0; i <= len; i++) {
-		/* i == len is the special case for the last line */
-		if (i == len || message[i] == '\n') {
-			tab_replace = 0;
-			while (message[j] == '\t' && j < i) {
-				tab_replace++;
-				j++;
-			}
-
-			XDrawString(dpy, win, gc, width + tab_size*tab_replace, height + 20*k, message + j, i - j);
-			while (i < len && message[i] == '\n') {
-				i++;
-				j = i;
-				k++;
-			}
-		}
-	}
-
-	/* xsi should not be NULL anyway if Xinerama is active, but to be safe */
-	if (XineramaIsActive(dpy) && xsi != NULL)
-			XFree(xsi);
-}
-
-
-
-static const char *
-gethash(void)
-{
-	const char *hash;
-	struct passwd *pw;
-
-	/* Check if the current user has a password entry */
-	errno = 0;
-	if (!(pw = getpwuid(getuid()))) {
-		if (errno)
-			die("instantlock: getpwuid: %s\n", strerror(errno));
-		else
-			die("instantlock: cannot retrieve password entry\n");
-	}
-	hash = pw->pw_passwd;
-
-#if HAVE_SHADOW_H
-	if (!strcmp(hash, "x")) {
-		struct spwd *sp;
-		if (!(sp = getspnam(pw->pw_name)))
-			die("instantlock: getspnam: cannot retrieve shadow entry. "
-			    "Make sure to suid or sgid instantlock.\n");
-		hash = sp->sp_pwdp;
-	}
-#else
-	if (!strcmp(hash, "*")) {
-#ifdef __OpenBSD__
-		if (!(pw = getpwuid_shadow(getuid())))
-			die("instantlock: getpwnam_shadow: cannot retrieve shadow entry. "
-			    "Make sure to suid or sgid instantlock.\n");
-		hash = pw->pw_passwd;
-#else
-		die("instantlock: getpwuid: cannot retrieve shadow entry. "
-		    "Make sure to suid or sgid instantlock.\n");
-#endif /* __OpenBSD__ */
-	}
-#endif /* HAVE_SHADOW_H */
-
-	return hash;
-}
-
-static void
-readpw(Display *dpy, struct xrandr *rr, struct lock **locks, int nscreens,
-       const char *hash)
-{
-	XRRScreenChangeNotifyEvent *rre;
-	char buf[32], passwd[256], *inputhash;
-	char tmpmessage[300] = "entering password :                       ";
-	int num, screen, running, failure, oldc;
-	unsigned int len, color;
-	int i;
-	KeySym ksym;
-	XEvent ev;
-
-	len = 0;
-	running = 1;
-	failure = 0;
-	oldc = INIT;
-
-	while (running && !XNextEvent(dpy, &ev)) {
-		if (ev.type == KeyPress) {
-			if (onebutton) {
-				running = 0;
-			}
-			explicit_bzero(&buf, sizeof(buf));
-			num = XLookupString(&ev.xkey, buf, sizeof(buf), &ksym, 0);
-			if (IsKeypadKey(ksym)) {
-				if (ksym == XK_KP_Enter)
-					ksym = XK_Return;
-				else if (ksym >= XK_KP_0 && ksym <= XK_KP_9)
-					ksym = (ksym - XK_KP_0) + XK_0;
-			}
-			if (IsFunctionKey(ksym) ||
-			    IsKeypadKey(ksym) ||
-			    IsMiscFunctionKey(ksym) ||
-			    IsPFKey(ksym) ||
-			    IsPrivateKeypadKey(ksym))
-				continue;
-			switch (ksym) {
-			case XK_Return:
-				if (onebutton) {
-					running = 0;
-					break;
-				}
-				passwd[len] = '\0';
-				errno = 0;
-				if (!(inputhash = crypt(passwd, hash)))
-					fprintf(stderr, "instantlock: crypt: %s\n", strerror(errno));
-				else
-					running = !!strcmp(inputhash, hash);
-				if (running) {
-					XBell(dpy, 100);
-					failure = 1;
-				}
-				explicit_bzero(&passwd, sizeof(passwd));
-				len = 0;
-				break;
-			case XK_Escape:
-				explicit_bzero(&passwd, sizeof(passwd));
-				len = 0;
-				break;
-			case XK_BackSpace:
-				if (len)
-					passwd[--len] = '\0';
-				break;
-			default:
-				if (num && !iscntrl((int)buf[0]) &&
-				    (len + num < sizeof(passwd))) {
-					memcpy(passwd + len, buf, num);
-					len += num;
-				}
-				break;
-			}
-			color = len ? INPUT : ((failure || failonclear) ? FAILED : INIT);
-			if (color == INPUT) {
-				message = tmpmessage;
-				for (int i = 0; i < strlen(passwd); i++)
-				{
-					tmpmessage[i + 20] = '.';
-					tmpmessage[i + 21] = '\0';
-				}
-			}
-			if (running /* && oldc != color */) {
-				for (screen = 0; screen < nscreens; screen++) {
-					XSetWindowBackground(dpy,
-					                     locks[screen]->win,
-					                     locks[screen]->colors[color]);
-					XClearWindow(dpy, locks[screen]->win);
-					writemessage(dpy, locks[screen]->win, screen);
-				}
-				oldc = color;
-			}
-		} else if (rr->active && ev.type == rr->evbase + RRScreenChangeNotify) {
-			rre = (XRRScreenChangeNotifyEvent*)&ev;
-			for (screen = 0; screen < nscreens; screen++) {
-				if (locks[screen]->win == rre->window) {
-					if (rre->rotation == RR_Rotate_90 ||
-					    rre->rotation == RR_Rotate_270)
-						XResizeWindow(dpy, locks[screen]->win,
-						              rre->height, rre->width);
-					else
-						XResizeWindow(dpy, locks[screen]->win,
-						              rre->width, rre->height);
-					XClearWindow(dpy, locks[screen]->win);
-					break;
-				}
-			}
-		} else {
-			for (screen = 0; screen < nscreens; screen++)
-				XRaiseWindow(dpy, locks[screen]->win);
-		}
-	}
-}
-
-static struct lock *
-lockscreen(Display *dpy, struct xrandr *rr, int screen)
-{
-	char curs[] = {0, 0, 0, 0, 0, 0, 0, 0};
-	int i, ptgrab, kbgrab;
-	struct lock *lock;
-	XColor color, dummy;
-	XSetWindowAttributes wa;
-	Cursor invisible;
-
-	if (dpy == NULL || screen < 0 || !(lock = malloc(sizeof(struct lock))))
-		return NULL;
-
-	lock->screen = screen;
-	lock->root = RootWindow(dpy, lock->screen);
-
-	for (i = 0; i < NUMCOLS; i++) {
-		XAllocNamedColor(dpy, DefaultColormap(dpy, lock->screen),
-		                 colorname[i], &color, &dummy);
-		lock->colors[i] = color.pixel;
-	}
-
-	/* init */
-	wa.override_redirect = 1;
-	wa.background_pixel = lock->colors[INIT];
-	lock->win = XCreateWindow(dpy, lock->root, 0, 0,
-	                          DisplayWidth(dpy, lock->screen),
-	                          DisplayHeight(dpy, lock->screen),
-	                          0, DefaultDepth(dpy, lock->screen),
-	                          CopyFromParent,
-	                          DefaultVisual(dpy, lock->screen),
-	                          CWOverrideRedirect | CWBackPixel, &wa);
-	lock->pmap = XCreateBitmapFromData(dpy, lock->win, curs, 8, 8);
-	invisible = XCreatePixmapCursor(dpy, lock->pmap, lock->pmap,
-	                                &color, &color, 0, 0);
-	XDefineCursor(dpy, lock->win, invisible);
-
-	/* Try to grab mouse pointer *and* keyboard for 600ms, else fail the lock */
-	for (i = 0, ptgrab = kbgrab = -1; i < 6; i++) {
-		if (ptgrab != GrabSuccess) {
-			ptgrab = XGrabPointer(dpy, lock->root, False,
-			                      ButtonPressMask | ButtonReleaseMask |
-			                      PointerMotionMask, GrabModeAsync,
-			                      GrabModeAsync, None, invisible, CurrentTime);
-		}
-		if (kbgrab != GrabSuccess) {
-			kbgrab = XGrabKeyboard(dpy, lock->root, True,
-			                       GrabModeAsync, GrabModeAsync, CurrentTime);
-		}
-
-		/* input is grabbed: we can lock the screen */
-		if (ptgrab == GrabSuccess && kbgrab == GrabSuccess) {
-			XMapRaised(dpy, lock->win);
-			if (rr->active)
-				XRRSelectInput(dpy, lock->win, RRScreenChangeNotifyMask);
-
-			XSelectInput(dpy, lock->root, SubstructureNotifyMask);
-			return lock;
-		}
-
-		/* retry on AlreadyGrabbed but fail on other errors */
-		if ((ptgrab != AlreadyGrabbed && ptgrab != GrabSuccess) ||
-		    (kbgrab != AlreadyGrabbed && kbgrab != GrabSuccess))
-			break;
-
-		usleep(100000);
-	}
-
-	/* we couldn't grab all input: fail out */
-	if (ptgrab != GrabSuccess)
-		fprintf(stderr, "instantlock: unable to grab mouse pointer for screen %d\n",
-		        screen);
-	if (kbgrab != GrabSuccess)
-		fprintf(stderr, "instantlock: unable to grab keyboard for screen %d\n",
-		        screen);
-	return NULL;
-}
-
-static void
-monitorreset(Display* dpy, CARD16 standby, CARD16 suspend, CARD16 off)
-{
-	DPMSSetTimeouts(dpy, standby, suspend, off);
-	DPMSForceLevel(dpy, DPMSModeOn);
-	XFlush(dpy);
-}
-
-static void
-usage(void)
-{
-	die("usage: instantlock [-v] [-f] [-m message] [cmd [arg ...]]\n");
-}
+#include "lock.h"
 
 int
 main(int argc, char **argv) {
-	struct xrandr rr;
-	struct lock **locks;
-	struct passwd *pwd;
-	struct group *grp;
-	uid_t duid;
-	gid_t dgid;
-	const char *hash;
 	Display *dpy;
-	int i, s, nlocks, nscreens;
-	CARD16 standby, suspend, off;
-	int count_fonts;
 	char **font_names;
+	int count_fonts;
+	int oneButton = onebutton;
+	int runCommand = 0;
+	const char* lock_message = message;
+	struct arg_lit *help, *version, *fonts, *anyKey;
+	struct arg_str *message, *command;
+	struct arg_end *end;
 
-	ARGBEGIN {
-	case 'v':
-		fprintf(stderr, "instantlock-"VERSION"\n");
-		return 0;
-	case 'm':
-		message = EARGF(usage());
-		break;
-	case 'o':
-		onebutton = 1;
-		message = "Press key to Unlock";
-		break;
-	case 'f':
+
+	void* argtable[] = {
+		help		= arg_litn(NULL, "help", 0, 1, "display this help and exit"),
+		version		= arg_litn("v", "version", 0, 1, "print instantlock version"),
+		fonts		= arg_litn("f", "fonts", 0, 1, "list available fonts"),
+		anyKey		= arg_litn("o", NULL, 0, 1, "press any key"),
+		message		= arg_str0("m", "message", "MESSAGE", "display the specified message"),
+		command		= arg_strn(NULL, NULL, "COMMAND [ARGS ...]",0,argc+2,NULL),
+		end			= arg_end(20)
+	};
+
+	int nerrors;
+	int exitcode = 0;
+	char progname[] = "instantlock";
+
+	/* verify the argtable[] entries were allocated sucessfully */
+    if (arg_nullcheck(argtable) != 0)
+	{
+		/* NULL entries were detected, some allocations must have failed */
+		printf("%s: insufficient memory\n",progname);
+		exitcode = 1;
+		goto exit;
+	}
+
+	nerrors = arg_parse(argc,argv,argtable);
+
+	/* special case: '--help' takes precedence over error reporting */
+    if (help->count > 0)
+    {
+        printf("Usage: %s", progname);
+        arg_print_syntax(stdout, argtable, "\n");
+        printf("Demonstrate command-line parsing in argtable3.\n\n");
+        arg_print_glossary(stdout, argtable, "  %-25s %s\n");
+		exitcode = 0;
+		goto exit;
+    }
+
+	if (command->count > 0)
+		runCommand = 1;
+
+	/* If the parser returned any errors then display them and exit */
+    if (nerrors > 0)
+    {
+        /* Display the error details contained in the arg_end struct.*/
+        arg_print_errors(stdout, end, progname);
+        printf("Try '%s --help' for more information.\n", progname);
+        exitcode = 1;
+        goto exit;
+    }
+	if (version->count > 0)
+	{
+		fputs("instantlock-"VERSION"\n", stderr);
+		exitcode = 0;
+		goto exit;
+	}
+	if (message->count > 0)
+	{
+		lock_message = message->sval[0];
+	}
+	if (anyKey->count > 0)
+	{
+		oneButton = 1;
+		lock_message = "Press key to Unlock";
+	}
+	if (fonts->count > 0)
+	{
 		if (!(dpy = XOpenDisplay(NULL)))
 			die("instantlock: cannot open display\n");
 		font_names = XListFonts(dpy, "*", 10000 /* list 10000 fonts*/, &count_fonts);
+		int i = 0;
 		for (i=0; i<count_fonts; i++) {
 			fprintf(stderr, "%s\n", *(font_names+i));
 		}
-		return 0;
-	default:
-		usage();
-	} ARGEND
-
-	/* validate drop-user and -group */
-	errno = 0;
-	if (!(pwd = getpwnam(user)))
-		die("instantlock: getpwnam %s: %s\n", user,
-		    errno ? strerror(errno) : "user entry not found");
-	duid = pwd->pw_uid;
-	errno = 0;
-	if (!(grp = getgrnam(group)))
-		die("instantlock: getgrnam %s: %s\n", group,
-		    errno ? strerror(errno) : "group entry not found");
-	dgid = grp->gr_gid;
-
-#ifdef __linux__
-	dontkillme();
-#endif
-
-	hash = gethash();
-	errno = 0;
-	if (!crypt("", hash))
-		die("instantlock: crypt: %s\n", strerror(errno));
-
-	if (!(dpy = XOpenDisplay(NULL)))
-		die("instantlock: cannot open display\n");
-
-	/* drop privileges */
-	if (setgroups(0, NULL) < 0)
-		die("instantlock: setgroups: %s\n", strerror(errno));
-	if (setgid(dgid) < 0)
-		die("instantlock: setgid: %s\n", strerror(errno));
-	if (setuid(duid) < 0)
-		die("instantlock: setuid: %s\n", strerror(errno));
-
-	/* check for Xrandr support */
-	rr.active = XRRQueryExtension(dpy, &rr.evbase, &rr.errbase);
-
-	/* get number of screens in display "dpy" and blank them */
-	nscreens = ScreenCount(dpy);
-	if (!(locks = calloc(nscreens, sizeof(struct lock *))))
-		die("instantlock: out of memory\n");
-	for (nlocks = 0, s = 0; s < nscreens; s++) {
-		if ((locks[s] = lockscreen(dpy, &rr, s)) != NULL) {
-			writemessage(dpy, locks[s]->win, s);
-			nlocks++;
-		} else {
-			break;
-		}
+		exitcode = 0;
+		goto exit;
 	}
-	XSync(dpy, 0);
-
-	/* did we manage to lock everything? */
-	if (nlocks != nscreens)
-		return 1;
-
-	/* DPMS-magic to disable the monitor */
-	if (!DPMSCapable(dpy))
-		die("instantlock: DPMSCapable failed\n");
-	if (!DPMSEnable(dpy))
-		die("instantlock: DPMSEnable failed\n");
-	if (!DPMSGetTimeouts(dpy, &standby, &suspend, &off))
-		die("instantlock: DPMSGetTimeouts failed\n");
-	if (!standby || !suspend || !off)
-		/* set values if there arent some */
-		standby = suspend = off = 300;
-
-	DPMSSetTimeouts(dpy, monitortime, monitortime, monitortime);
-	XFlush(dpy);
-
-	/* run post-lock command */
-	if (argc > 0) {
-		switch (fork()) {
-		case -1:
-			die("instantlock: fork failed: %s\n", strerror(errno));
-		case 0:
-			monitorreset(dpy, standby, suspend, off);
-
-			if (close(ConnectionNumber(dpy)) < 0)
-				die("instantlock: close: %s\n", strerror(errno));
-			execvp(argv[0], argv);
-			fprintf(stderr, "instantlock: execvp %s: %s\n", argv[0], strerror(errno));
-			_exit(1);
-		}
+	if (command->count > 0) {
+		runCommand = 1;
 	}
 
-	/* everything is now blank. Wait for the correct password */
-	readpw(dpy, &rr, locks, nscreens, hash);
+	fprintf(stderr, "lock message : %s\n", lock_message);
+	fprintf(stderr, "oneButton : %d\n", oneButton);
+	fprintf(stderr, "runCommand : %d\n", runCommand);
 
-	/* reset DPMS values to inital ones */
-	monitorreset(dpy, standby, suspend, off);
+	lock(lock_message, oneButton, runCommand, argv);
 
-	return 0;
+exit:
+	arg_freetable(argtable, sizeof(argtable) / sizeof(argtable[0]));
+	return exitcode;
 }
